:hammer: Refactor grpc code into seperate modules
