+++ croncat/src/grpc/mod.rs
@@ -1,0 +1,11 @@
+//!
+//! Use the [cosmos_sdk_proto](https://crates.io/crates/cosmos-sdk-proto) library to create clients for GRPC node requests.
+//!
+
+pub mod querier;
+pub mod service;
+pub mod signer;
+
+pub use querier::GrpcQuerier;
+pub use service::GrpcClientService;
+pub use signer::GrpcSigner;
